chore: add PostHog analytics
PostHog analytics is now available both based on their NPM module `posthog-js` as well as by simply fetching their basic analytics script. Currently, the module-based tracking is in use as this is expected to be faster after the first page has fetched the JavaScript bundle and subsequent pages only need to capture the events.